refactor(challenge-record): 탭 패널에 해당하는 컴포넌트들 파일 경로 변경
